requirements.txt
# manual
pydantic>=2.0
hydra_colorlog
exact_cover

# from pip list
absl-py==2.1.0
accelerate==0.29.3
ai2thor==5.0.0
aiohttp==3.8.5
aiosignal==1.3.1
algo==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
aws-requests-auth==0.4.3
backcall==0.2.0
bitsandbytes==0.39.1
black==21.4b2
botocore==1.34.64
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
cloudpickle==3.0.0
cmake==3.25.0
contourpy==1.0.5
cryptography==41.0.2
cycler==0.11.0
datasets==2.18.0
decorator==5.1.1
# detectron2==0.6 - install from source
dgl==2.1.0
dill==0.3.8
diskcache==5.6.1
distro==1.9.0
docstring_parser==0.16
editdistance==0.8.1
einops==0.6.1
exceptiongroup==1.2.0
executing==1.2.0
fastapi==0.110.0
filelock==3.12.2
FlagEmbedding==1.2.8
Flask==1.1.2
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2023.6.0
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.29.0
google-auth-oauthlib==1.0.0
gptcache==0.1.37
grpcio==1.62.1
guidance==0.0.64
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
# habitat==0.2.1 - install from source
# habitat-sim==0.2.1 - install from source
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
hydra-core==1.3.2
idna==3.4
ifcfg==0.24
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
importlib-resources==5.2.0
iopath==0.1.9
ipdb==0.13.13
ipython==8.12.2
itsdangerous==1.1.0
jedi==0.18.2
Jinja2==2.11.3
jmespath==1.0.1
joblib==1.3.2
jsonargparse==4.27.6
kiwisolver==1.4.4
lazy_loader==0.3
# lightning @ git+https://github.com/Lightning-AI/lightning@master
lit==15.0.7
littleutils==0.2.2
llama_cpp_python==0.2.56
llvmlite==0.41.1
Markdown==3.6
MarkupSafe==1.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.2.1
msal==1.23.0
msgpack==1.0.8
multidict==6.0.4
multiprocess==0.70.16
munkres==1.1.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1.1
numba==0.58.1
numpy==1.24.3
numpy-quaternion==2023.0.3
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.13.3
opencv-python==4.8.0.74
ordered-set==4.1.0
outdated==0.2.2
packaging==23.0
pandas==2.0.3
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==24.0
platformdirs==3.9.1
ply==3.11
portalocker==2.8.2
progressbar2==4.2.0
prompt-toolkit==3.0.39
protobuf==5.26.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydot==2.0.0
pyformlang==1.0.7
Pygments==2.15.1
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
python-dateutil==2.8.2
python-utils==3.7.0
python-xlib==0.33
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.4.0
revtok==0.0.3
rsa==4.9
safetensors==0.4.2
scikit-fmm==2023.4.2
scikit-image==0.21.0
scikit-learn==0.22.2.post1
scipy==1.10.1
sentence-transformers==2.6.1
sentencepiece==0.1.99
setuptools==67.8.0
sip==6.6.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
socksio==1.0.0
stack-data==0.6.2
starlette==0.36.3
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2023.7.10
tiktoken==0.4.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.38.2
transformers-stream-generator==0.0.5
triton==2.0.0
typeshed_client==2.5.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.28.0
vocab==0.0.5
wcwidth==0.2.6
Werkzeug==1.0.1
wheel==0.38.4
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.2
zipp==3.11.0
zstandard==0.22.0
